chore: update renamed file imports in other files.
